Extract constants and json-ld from data
