+++ src/Entities/Search.tsx
@@ -3,7 +3,8 @@ import AsyncSelect from "react-select/async";
 import { Theme } from "react-select";
 import "./Search.css";
 
-import { getAutoCompletionSuggestions, JsonLdReference } from "./data";
+import { getAutoCompletionSuggestions } from "./data";
+import { JsonLdReference } from "./json-ld";
 import { PRIMARY, LIST_ITEM_ACTIVE, LIST_ITEM_HOVER } from "./colors";
 import EntityValue from "./EntityValue";
