+++ src/Entities/constants.ts
@@ -1,0 +1,27 @@
+// Namespaces
+export const RDFS = "http://www.w3.org/2000/01/rdf-schema#";
+export const RDF = "http://www.w3.org/1999/02/22-rdf-syntax-ns#";
+export const XSD = "http://www.w3.org/2001/XMLSchema#";
+export const OWL = "http://www.w3.org/2002/07/owl#";
+
+// RDF
+export const RDF_TYPE = RDF + "type";
+export const RDF_PROPERTY = RDF + "Property";
+
+// RDFS
+export const RDFS_LABEL = RDFS + "label";
+export const RDFS_CLASS = RDFS + "Class";
+export const RDFS_SUB_CLASS_OF = RDFS + "subClassOf";
+export const RDFS_COMMENT = RDFS + "comment";
+
+// XSD
+export const XSD_STRING = XSD + "string";
+
+// JSON LD
+export const JSON_LD_TYPE = "@type";
+
+// OWL
+export const OWL_CLASS = OWL + "Class";
+export const OWL_OBJECT_PROPERTY = OWL + "ObjectProperty";
+export const OWL_DATA_PROPERTY = OWL + "DataProperty";
+export const OWL_ANNOTATION_PROPERTY = OWL + "AnnotationProperty";
